Refactor Continue Watching data layer
* **refactor: Extract ProgressContinueWatchingFetcher**
    * Move the multi-step progress resolution logic from `ProgressContinueWatchingStore` into a dedicated `ProgressContinueWatchingFetcher` class.
    * This improves separation of concerns by isolating network resolution logic from Store5 management.
* **refactor: Standardize fetcher execution**
    * Update both `NitroContinueWatchingFetcher` and `ProgressContinueWatchingFetcher` to use the `invoke` operator instead of a named `run()` method.
* **refactor: Simplify ProgressContinueWatchingStore**
    * Reduce the dependency surface of `ProgressContinueWatchingStore` by delegating network tasks to the new fetcher.
* **docs: Enhance FetcherSkipSignal**
    * Add KDoc to `FetcherSkipSignal` explaining how it interacts with the Store5 writer and repository to preserve local data during upstream failures.
* **test: Update unit tests**
    * Update `ContinueWatchingFetcherTest`, `DefaultContinueWatchingRepositoryTest`, and `NitroContinueWatchingFetcherTest` to reflect the new fetcher architecture and method signatures.
